Move all bot commands to a new "commands" module
